profile: Move fixedvec into its own header file
